requirements.txt
gradio>=4.0.0
Pillow>=10.0.0
reportlab>=4.0.0
numpy>=1.24.0
